# documentation/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "description": "",
        "title": "Road Protect API",
        "version": "1.4.1"
    },
    "paths": {
...
        "/api/v1/contract/reference/{reference}/end-date": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "reference",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateContractEndDateDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Contract"
                        }
                    },
                    "400": {
                        "description": " * E051: More than one contract found with the same reference (reference), please ensure your contract reference is unique\n * E052: No contract found relating to your current account with that reference"
                    },
                    "500": {
                        "description": "Failed to update contract, please contact support"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Contracts"
                ],
                "operationId": "ContractController_updateContractEndDateByReference",
                "summary": "Update contract end date by reference"
            }
        },
        "/api/v1/contract/vehicle/{vehicleId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "vehicleId",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/Contract"
                            },
                            "type": "array"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Contracts"
                ],
                "operationId": "ContractController_getContractsForVehicle",
                "summary": "Get contracts for a vehicle by vehicleId"
            }
        },
        "/api/v1/contract/{contractId}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "contractId",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Contract"
                        }
                    },
                    "400": {
                        "description": "E035: Could not find the Contract to delete"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Contracts"
                ],
                "operationId": "ContractController_deleteContract",
                "summary": "Delete contract by contractId"
            },
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "contractId",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Contract"
                        }
                    },
                    "400": {
                        "description": "E126: Could not find the Contract with id: contractId"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Contracts"
                ],
                "operationId": "ContractController_getContract",
                "summary": "Get contract (either lease / ownership) by contractId"
            }
        },
        "/api/v1/contract/{contractId}/end-date": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "contractId",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateContractEndDateDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Contract"
                        }
                    },
                    "400": {
                        "description": "E050: Contract with id contractId not found"
                    },
                    "500": {
                        "description": "Failed to update contract, please contact support"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Contracts"
                ],
                "operationId": "ContractController_updateContractEndDate",
                "summary": "Update contract end date"
            }
        },
        "/api/v1/contract/{contractId}/reference": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "contractId",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateContractReferenceDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Contract"
                        }
                    },
                    "400": {
                        "description": "E050: Contract with id contractId not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Contracts"
                ],
                "description": "Your account has to be on the contract you are editing",
                "operationId": "ContractController_updateContractReferenceById",
                "summary": "Update contract reference by contractId"
            }
        },
...
        "/api/v1/vehicle/{vehicleId}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "vehicleId",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Vehicle"
                        }
                    },
                    "400": {
                        "description": "E157: Could not find the Vehicle to delete"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Vehicles"
                ],
                "operationId": "VehicleController_deleteVehicle",
                "summary": "Delete vehicle"
            },
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "vehicleId",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Vehicle"
                        }
                    },
                    "400": {
                        "description": "E049: Could not find Vehicle with id: vehicleId"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Vehicles"
                ],
                "operationId": "VehicleController_getVehicle",
                "summary": "Get vehicle"
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "vehicleId",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateVehicleDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Vehicle"
                        }
                    },
                    "400": {
                        "description": "Vehicle not found"
                    },
                    "500": {
                        "description": "Failed to update the vehicle - please contact the developers"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Vehicles"
                ],
                "operationId": "VehicleController_updateVehicle",
                "summary": "Edit vehicle"
            }
        }
    },
    "definitions": {
...
        "Contract": {
            "title": "Contract",
            "properties": {
                "contractId": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "document": {
                    "description": "Document",
                    "type": "object"
                },
                "endDate": {
                    "type": "string"
                },
                "infringements": {
                    "description": "Infringement[]",
                    "type": "object"
                },
                "owner": {
                    "description": "Account",
                    "type": "object"
                },
                "redirectionTargetAddress": {
                    "description": "This is a calculated field",
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "Upcoming",
                        "Expired",
                        "Expiring soon",
                        "Valid"
                    ],
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicle": {
                    "description": "Vehicle",
                    "type": "object"
                }
            },
            "required": [
                "createdAt",
                "updatedAt",
                "contractId",
                "type",
                "startDate",
                "endDate",
                "status",
                "reference",
                "document",
                "owner",
                "vehicle",
                "infringements",
                "redirectionTargetAddress"
            ],
            "type": "object"
        },
...
        "UpdateContractEndDateDto": {
            "title": "UpdateContractEndDateDto",
            "properties": {
                "endDate": {
                    "description": "The new contract end date",
                    "example": "2021-07-08T08:17:35.782Z",
                    "type": "string"
                },
                "startDate": {
                    "description": "The new contract start date",
                    "example": "2021-06-08T08:17:35.781Z",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "UpdateContractReferenceDto": {
            "title": "UpdateContractReferenceDto",
            "properties": {
                "reference": {
                    "description": "The new reference for this contract",
                    "example": "My internal reference 15192",
                    "type": "string"
                }
            },
            "required": [
                "reference"
            ],
            "type": "object"
        },
...
        "UpdateVehicleDto": {
            "title": "UpdateVehicleDto",
            "properties": {
                "category": {
                    "description": "The vehicle category class",
                    "example": "car 1-5",
                    "type": "string"
                },
                "color": {
                    "description": "The vehicle color",
                    "example": "gold",
                    "type": "string"
                },
                "manufacturer": {
                    "description": "The vehicle manufacturer",
                    "example": "Ford",
                    "type": "string"
                },
                "model": {
                    "description": "The vehicle model",
                    "example": "F40",
                    "type": "string"
                },
                "registration": {
                    "description": "The vehicle registration number",
                    "example": "51729193",
                    "type": "string"
                },
                "type": {
                    "description": "The type of vehicle in the fleet",
                    "enum": [
                        "Private",
                        "Truck"
                    ],
                    "example": "Private",
                    "type": "string"
                },
                "weight": {
                    "description": "The vehicle weight in tonnes",
                    "example": "1.9",
                    "type": "number"
                },
                "year": {
                    "description": "The year the vehicle was manufactured",
                    "example": "2012",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "Vehicle": {
            "title": "Vehicle",
            "properties": {
                "autoAssignTo": {
                    "enum": [
                        "Owner",
                        "User"
                    ],
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "contracts": {
                    "description": "Contract[]",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentLeaseContract": {
                    "description": "LeaseContract",
                    "type": "object"
                },
                "currentOwnershipContract": {
                    "description": "OwnershipContract",
                    "type": "object"
                },
                "infringements": {
                    "description": "Infringement",
                    "type": "object"
                },
                "locationRecords": {
                    "description": "IturanLocationRecord[]",
                    "type": "object"
                },
                "logs": {
                    "description": "Log[]",
                    "type": "object"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "modelYear": {
                    "type": "string"
                },
                "registration": {
                    "type": "string"
                },
                "type": {
                    "default": "Private",
                    "enum": [
                        "Private",
                        "Truck"
                    ],
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicleId": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                }
            },
            "required": [
                "createdAt",
                "updatedAt",
                "vehicleId",
                "registration",
                "manufacturer",
                "model",
                "modelYear",
                "color",
                "category",
                "weight",
                "type",
                "contracts",
                "currentLeaseContract",
                "currentOwnershipContract",
                "infringements",
                "autoAssignTo",
                "logs",
                "locationRecords"
            ],
            "type": "object"
        }
    },
    "securityDefinitions": {
        "bearer": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
...
}